Added callback to allow output to the new program 'console' text area.
Added CaptureCamera() and RestoreCamera() to (hopefully) allow the user to take screenshots from the same viewing angle every time. However, Dr. Ray suggested instead, to make two cameras, one for exploring and one that would generally stay static that could be switched to when the user requests a screenshot.

RenderFittedEllipsoid(): now takes user input to decide which fitting method to use
                         checks for the minimum of 9 points to fit an ellipsoid, warns the user if not

Removed the text actor displayed alongside the ellipsoid.

CalculateOrientations(): Added to allow the user to get stat information on the angles of orientation of bacteria

ColorByOrientations(): Makes the diffuse color RGB components equal to the dot product between the vector 
                        representing the bacterium and the 3-space basis vectors. Currently 3 user-selectable 
                        color schemes are selectable (different orders of x/y/z -> r/g/b)

Removed the MoveEllipsoid() and MoveEllipsoidText() methods since the fitting operates correctly now.